refactorn extern vs 'intern' use of bioservices modules
